+++ include/build_version.h
@@ -8,10 +8,10 @@
 #ifndef BUILD_VERSION_H
 #define BUILD_VERSION_H
 
-#define BUILD_NUMBER 996
-#define BUILD_TIMESTAMP "2026-01-06 22:00:36"
+#define BUILD_NUMBER 1002
+#define BUILD_TIMESTAMP "2026-01-06 22:31:13"
 #define BUILD_DATE "20260106"
-#define GIT_HASH "98df5ac"
+#define GIT_HASH "f7b061e"
 #define GIT_BRANCH "main"
 
 #endif // BUILD_VERSION_H